requirements.txt
streamlit==1.48.1
moviepy==2.0.0
Pillow>=10.2.0
numpy
imageio
imageio-ffmpeg
